Added VolumeConfiguration framework for mod/user predefined tank sets.
Added AbstractResourceTank to facilitate the use of CC as a patch mod via
DoCostPatch flag. Provided that flag tank modules calculate the cost offset so
that the patched part retains its cost. This only works the first time a tank is
initialized (presumably in editor).
Added Include/ExcludeTankTypes options to tank modules to restrict tank types
accessible in a particular part.
Changed TankCostPerVolume to PerSurface which is calculated as 6*V^(2/3) (i.e.
surface of cube of volume V).
Fixed maxTemp in HangarResourceContainer parts.